+++ src/Bcgov/EmergencyInfo/blocks/post-social-share/index.js
@@ -1,0 +1,12 @@
+/**
+ * WordPress dependencies
+ */
+import { registerBlockType } from '@wordpress/blocks';
+import metadata from './block.json';
+import { edit } from './edit';
+import './style.scss';
+
+registerBlockType(metadata, {
+	edit,
+	save: () => null,
+});